# library_api_gateway/api/docs/swagger.yaml
definitions:
...
  models.AllStuBookList:
    properties:
      bk_authorname:
        type: string
      bk_id:
        type: string
      bk_name:
        type: string
      bk_numsbook:
        type: integer
      book_list_id:
        type: integer
      deadline:
        type: string
      given_date:
        type: string
      status:
        type: string
      student_id:
        type: string
    type: object
...
  models.GetAllStudentLibraryResponse:
    properties:
      count:
        type: integer
      student_library:
        items:
          $ref: '#/definitions/models.Student'
        type: array
    type: object
  models.GetAllStudentResponse:
    properties:
      count:
        type: integer
      students:
        items:
          $ref: '#/definitions/models.Student'
        type: array
    type: object
...
  models.GetStudentLibraryInfo:
    properties:
      student_library:
        items:
          $ref: '#/definitions/models.Student'
        type: array
    type: object
...
  models.ResponseModel:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
  models.StuBookList:
    properties:
      bk_id:
        type: string
      deadline:
        type: string
      student_id:
        type: string
    type: object
  models.Student:
    properties:
      all_booklist:
        items:
          $ref: '#/definitions/models.AllStuBookList'
        type: array
      student_course:
        type: integer
      student_faculty:
        type: string
      student_firstname:
        type: string
      student_id:
        type: string
      student_lastname:
        type: string
      student_phone1:
        type: string
      student_phone2:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/stubooklist:
    get:
      consumes:
      - application/json
      description: Get Student BookList
      operationId: get-student-booklist
      parameters:
      - description: student_id
        in: query
        name: student_id
        type: string
      - description: student_firstname
        in: query
        name: student_firstname
        type: string
      - description: student_lastname
        in: query
        name: student_lastname
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.GetStudentLibraryInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get student booklist
      tags:
      - studentbooklist
    post:
      consumes:
      - application/json
      description: Create Student BookList
      operationId: create-student-booklist
      parameters:
      - description: stubook_list
        in: body
        name: stubook_list
        required: true
        schema:
          $ref: '#/definitions/models.StuBookList'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: create student booklist
      tags:
      - studentbooklist
    put:
      consumes:
      - application/json
      description: Return Student BookList
      operationId: return-student-booklist
      parameters:
      - description: bk_id
        in: query
        name: bk_id
        required: true
        type: string
      - description: stubooklist_id
        in: query
        name: stubooklist_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: return student booklist
      tags:
      - studentbooklist
  /v1/stubooklists:
    get:
      consumes:
      - application/json
      description: Get Student BookLists
      operationId: get-student-booklists
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.GetAllStudentLibraryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get student booklists
      tags:
      - studentbooklist
  /v1/student:
    delete:
      consumes:
      - application/json
      description: Delete Student by given ID
      operationId: delete_student
      parameters:
      - description: student_id
        in: query
        name: student_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Delete Student
      tags:
      - student
    get:
      consumes:
      - application/json
      description: Get Student
      operationId: get_student
      parameters:
      - description: student_id
        in: query
        name: student_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            type: Object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Get Student
      tags:
      - student
    post:
      consumes:
      - application/json
      description: Create Student
      operationId: create_student
      parameters:
      - description: student
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Create Student
      tags:
      - student
    put:
      consumes:
      - application/json
      description: Update Student by ID
      operationId: update_student
      parameters:
      - description: student
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Update Student
      tags:
      - student
  /v1/students:
    get:
      consumes:
      - application/json
      description: Get All Student
      operationId: get_all_student
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.GetAllStudentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Get All Student
      tags:
      - student
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"